backend/requirements.txt
# FastAPI 核心
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# 数据库
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# 文档处理
python-docx>=1.1.0

# AI服务
dashscope>=1.17.0

# NLP
nltk>=3.8.0

# 数据验证
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 工具
python-dotenv>=1.0.0
httpx>=0.26.0

# 开发工具
pytest>=7.4.0
pytest-asyncio>=0.23.0
